Add support for 3D sprites

+++ examples/common/mod.rs
@@ -4,8 +4,8 @@ use rand::Rng;
 const DEFAULT_WINDOW_WIDTH: f32 = 1280.0;
 const DEFAULT_WINDOW_HEIGHT: f32 = 720.0;
 
-/// Returns the position of the nth item in a grid
-pub fn grid_position(columns: u32, rows: u32, n: u32) -> Vec3 {
+/// Returns the screen-space position of the nth item in a grid
+pub fn grid_position(columns: u32, rows: u32, n: usize) -> Vec3 {
     const MARGIN: f32 = 100.0;
 
     let width = DEFAULT_WINDOW_WIDTH - MARGIN * 2.0;
@@ -14,8 +14,8 @@ pub fn grid_position(columns: u32, rows: u32, n: u32) -> Vec3 {
     let xgap = width / columns.saturating_sub(1) as f32;
     let ygap = height / rows.saturating_sub(1) as f32;
 
-    let x = (n % columns) as f32;
-    let y = (n / columns) as f32;
+    let x = (n as u32 % columns) as f32;
+    let y = (n as u32 / columns) as f32;
 
     Vec3::new(
         x * xgap - width / 2.0,
